The old Index class has now been completely phased out, and is replaced by either vector<int> (for array indices) or vector<unsigned int> (for dimensions).
A first sweep through the source to remove long integers has been done.
Many integers and integer arguments are now signed, when necessary.

+++ src/include/Console.h
@@ -108,7 +108,7 @@ public:
    *
    * @returns true on success, false on failure.
    */ 
-  bool update (unsigned long n);
+  bool update (unsigned int n);
   /**
    * Sets a trace monitor for a subset of the given variable
    *
@@ -150,7 +150,7 @@ public:
   /**
    * Returns the iteration number of the model.
    */
-  unsigned long iter(unsigned int chain) const;
+  unsigned int iter(unsigned int chain) const;
   /**
    * Dump the contants of monitored node in CODA format
    * The parameters name and range must correspond to a previous
